+++ test/integration/tests/3997-coverage-with-cabal-3/files/src/Lib.hs
@@ -1,0 +1,4 @@
+module Lib (foo) where
+
+foo :: Int
+foo = 1